@@ -11,19 +11,22 @@
   },
   "dependencies": {
     "blacklist": "^1.1.2",
-    "classnames": "^2.1.3"
+    "classnames": "^2.1.3",
+    "history": "^1.12.0"
   },
   "devDependencies": {
     "babel-eslint": "^4.1.3",
     "eslint": "^1.4.3",
     "eslint-plugin-react": "^3.4.1",
     "gulp": "^3.9.0",
-    "react": "^0.13.3",
+    "react": ">=0.14.0-rc1",
+    "react-dom": ">=0.14.0-rc1",
+    "react-addons-css-transition-group": "^0.14.0-rc1",
     "react-component-gulp-tasks": "^0.7.1",
-    "react-router": "^0.13.3"
+    "react-router": "1.0.0-rc1"
   },
   "peerDependencies": {
-    "react": "^0.13.0 || ^0.14.0-beta"
+    "react": "^0.13.0 || 0.14.0-rc1"
   },
   "browserify-shim": {
     "react": "global:React"
