@@ -1,4 +1,4 @@
-var React = require('react/addons');
+var React = require('react');
 
 module.exports = React.createClass({
 	displayName: 'BlankState',
