@@ -1,6 +1,6 @@
 var blacklist = require('blacklist');
 var classnames = require('classnames');
-var React = require('react/addons');
+var React = require('react');
 
 module.exports = React.createClass({
 	displayName: 'Form',
